requirements.txt
# Core dependencies
python-dateutil>=2.8.2
pytz>=2024.1

# Web scraping and feeds
feedparser>=6.0.10
beautifulsoup4>=4.12.0
requests>=2.31.0
aiohttp>=3.9.0
lxml>=5.0.0

# AI/ML
google-generativeai>=0.3.0
sentence-transformers>=2.2.2
torch>=2.0.0
transformers>=4.36.0
spacy>=3.7.0

# Database
psycopg2-binary>=2.9.9
redis>=5.0.0
sqlalchemy>=2.0.0
alembic>=1.13.0

# Vector search
qdrant-client>=1.7.0
numpy>=1.24.0
scikit-learn>=1.3.0

# NLP and text processing
langdetect>=1.0.9
nltk>=3.8.1
textstat>=0.7.3

# Data processing
pandas>=2.0.0
pydantic>=2.5.0
pyyaml>=6.0

# Web framework (for API)
fastapi>=0.108.0
uvicorn>=0.25.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.7.0
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.12.0
flake8>=6.1.0
mypy>=1.8.0
pre-commit>=3.6.0